requirements.txt
atomicwrites==1.3.0
attrs==18.2.0
CacheControl==0.12.5
cachetools==3.1.0
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
dj-database-url==0.5.0
Django==2.1.7
django-autofixture==0.12.1
django-cors-headers==2.5.0
django-crispy-forms==1.7.2
django-extensions==2.1.6
django-filter==2.1.0
django-heroku==0.3.1
django-rest-auth==0.9.3
djangorestframework==3.9.1
djangorestframework-firebase==0.2.0
drf-firebase-auth==0.0.4
export==0.1.2
firebase-admin==2.16.0
gcloud==0.17.0
google-api-core==1.8.0
google-auth==1.6.3
google-cloud-core==0.29.1
google-cloud-firestore==0.31.0
google-cloud-storage==1.14.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.8
grpcio==1.19.0
gunicorn==19.9.0
httplib2==0.12.1
idna==2.8
jws==0.1.3
Markdown==3.0.1
more-itertools==6.0.0
msgpack==0.6.1
oauth2client==3.0.0
Pillow==5.4.1
pluggy==0.9.0
protobuf==3.7.0rc2
psycopg2==2.7.7
psycopg2-binary==2.7.7
py==1.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycryptodome==3.7.3
pytest==4.3.0
python-firebase==1.2
python-jwt==2.0.1
pytz==2018.9
requests==2.21.0
requests-toolbelt==0.7.0
rsa==4.0
six==1.12.0
urllib3==1.24.1
whitenoise==4.1.2
django-import-export